+++ NEWS.md
@@ -1,4 +1,18 @@
-# brms 2.12.0++
+# brms 2.13.0++
+
+### Bug Fixes
+
+* Fix generated Stan Code of models with improper global priors and
+`constant` priors on some coefficients thanks to Frank Weber. (#919)
+
+### Other Changes
+
+* Adjust behavior of the `rate` addition term so that it also
+affects the `shape` parameter in `negbinomial` models thanks to
+Edward Abraham. (#915)
+
+
+# brms 2.13.0
 
 ### New Features